+++ R/truncate_interlace.R
@@ -103,13 +103,20 @@ truncate_interlace <- function(primary, secondary = NULL) {
                       (.data$valid_until > .data$valid_from.y)  |   # of the primary data.
                         is.na(.data$valid_until) |
                         is.na(.data$valid_until.y)) |>
-        dplyr::mutate("valid_from"  = pmax(.data$valid_from,  .data$valid_from.y,  na.rm = TRUE),
-                      "valid_until" = pmin(.data$valid_until, .data$valid_until.y, na.rm = TRUE)) |>
+        dplyr::mutate(
+          "valid_from"  = ifelse(.data$valid_from  >= .data$valid_from.y,  .data$valid_from, .data$valid_from.y),       # nolint: ifelse_censor_linter
+          "valid_until" = ifelse(.data$valid_until <= .data$valid_until.y, .data$valid_from, .data$valid_from.y)        # nolint: ifelse_censor_linter
+        ) |>
         dplyr::select(-tidyselect::ends_with(".y"))
     })
 
   # With the secondary data truncated, we can interlace and return
-  out <- SCDB::interlace_sql(secondary_truncated, by = purrr::pluck(primary_keys, 1))
+  if (packageVersion("SCDB") < "0.4.0") {
+    interlace <- SCDB::interlace_sql
+  } else {
+    interlace <- utils::getFromNamespace("interlace", "SCDB")
+  }
+  out <- interlace(secondary_truncated, by = purrr::pluck(primary_keys, 1))
 
   return(out)
 }