+++ tests/testthat/test-DiseasystoreBase.R
@@ -1,5 +1,6 @@
 # Set testing options
 withr::local_options("diseasystore.target_schema" = target_schema_1)
+withr::local_options("diseasystore.lock_wait_max" = 1 * 60) # 1 minute during tests
 
 test_that("DiseasystoreBase works", {
 
@@ -103,6 +104,7 @@ test_that("DiseasystoreBase works", {
 
 
 test_that("$get_feature verbosity works", {
+  skip_if_not_installed("RSQLite")
 
   # Create a dummy DiseasystoreBase with a mtcars FeatureHandler
   DiseasystoreDummy <- R6::R6Class(                                                                                     # nolint: object_name_linter
@@ -146,6 +148,7 @@ test_that("$get_feature verbosity works", {
 
 
 test_that("DiseasystoreBase$determine_new_ranges works", {
+  skip_if_not_installed("RSQLite")
 
   start_date <- as.Date("2020-01-01")
   end_date   <- as.Date("2020-03-01")
@@ -153,7 +156,13 @@ test_that("DiseasystoreBase$determine_new_ranges works", {
 
   conn <- DBI::dbConnect(RSQLite::SQLite())
   ds <- DiseasystoreBase$new(source_conn = "", target_conn = conn)
-  logs <- SCDB::create_logs_if_missing(paste(target_schema_1, "logs", sep = "."), conn)
+
+  if (packageVersion("SCDB") < "0.4.0") {
+    logs <- SCDB::create_logs_if_missing(paste(target_schema_1, "logs", sep = "."), conn)
+  } else {
+    logs <- SCDB::create_logs_if_missing(conn = conn, log_table = paste(target_schema_1, "logs", sep = "."))
+  }
+
 
   dplyr::rows_append(
     logs,